requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.5
django-multiupload==0.6.1
djangorestframework==3.14.0
Pillow==10.0.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
